replace next and previous lemon jelly
